eda_automator/requirements/image_generation.txt
# Requisitos para la generación de imágenes en EDA Automator
#
# Este archivo contiene las dependencias necesarias para generar reportes como imágenes
# desde HTML usando diversos métodos.
#
# Instalación:
# pip install -r image_generation.txt

# Método imgkit (requiere wkhtmltopdf instalado https://wkhtmltopdf.org/downloads.html)
imgkit>=1.2.2

# Método Selenium (requiere un navegador Chrome)
selenium>=4.1.0
webdriver-manager>=3.5.2

# Método WeasyPrint (genera PDF primero y luego convierte a imagen)
weasyprint>=53.0
pdf2image>=1.16.0
Pillow>=8.3.1 